refactor 2.0.1 block of changelog
Fixed blatantly incorrect date

Including items that e.g. describe changes to the changelog just bloats the changelog. It's supposed to be the point of reference for consumer-facing changes IIRC. Before 2.0, no changelog talked about tooling that didn't affect the code or tests either. (We can also remove the mentions of tests if needed.)